[BE][S13P21D108-107] refactor: API 호출 실패 시에도 Transaction 남기도록 수정

See merge request s13-fintech-finance-sub1/S13P21D108!60

+++ backend/moyeo/src/main/java/com/mo/moyeo/domain/transaction/bank/dto/BankTransferDTO.java
@@ -11,7 +11,7 @@ public class BankTransferDTO {
 
     private String depositAccountNo;
     private String depositTransactionSummary;
-    private Long transactionBalance;
+    private Double transactionBalance;
     private String withdrawalAccountNo;
     private String withdrawalTransactionSummary;
     private String userKey;
